abstracted todo removal animation
